add LineRange.js, add useEffect to LineTime.js
